Conflicts:
	src/bindings/ScriptDev2
	src/game/ScriptMgr.cpp

+++ src/game/ScriptMgr.cpp
@@ -34,6 +34,7 @@ CUSTUM STUFF BEGIN
 /*
 CUSTUM STUFF END
 */
+#include "SQLStorages.h"
 #include "revision_nr.h"
 
 ScriptMapMapName sQuestEndScripts;
@@ -49,7 +50,6 @@ INSTANTIATE_SINGLETON_1(ScriptMgr);
 
 ScriptMgr::ScriptMgr() :
     m_hScriptLib(NULL),
-    
     m_scheduledScripts(0),
 
     m_pOnInitScriptLibrary(NULL),
@@ -604,7 +604,7 @@ void ScriptMgr::LoadScripts(ScriptMapMapName& scripts, const char* tablename)
                         if (SpellEntry const* spell = sSpellStore.LookupEntry(i))
                             for (int j = 0; j < MAX_EFFECT_INDEX; ++j)
                             {
-                                if (spell->Effect[j] == SPELL_EFFECT_SEND_TAXI && spell->EffectMiscValue[j] == tmp.sendTaxiPath.taxiPathId)
+                                if (spell->Effect[j] == SPELL_EFFECT_SEND_TAXI && spell->EffectMiscValue[j] == int32(tmp.sendTaxiPath.taxiPathId))
                                 {
                                     taxiSpell = i;
                                     break;
